+++ client/Scripts/View/HeroEntity.cs
@@ -42,7 +42,7 @@ public partial class HeroEntity : TeamEntityViewNode, IPlayerView, IAttackableVi
     // TODO: attack animation / particles
   }
 
-  public void OnHitVfx(int damage, Vector3 attackerPosition) {
+  public void OnHitVfx(float damage, Vector3 attackerPosition) {
     // TODO: hit reaction / particles
   }
