+++ client/Scripts/View/IAttackableView.cs
@@ -4,5 +4,5 @@ namespace Meesles.Avalon;
 
 public interface IAttackableView {
   void OnAttackVfx(Vector3 targetPosition);
-  void OnHitVfx(int damage, Vector3 attackerPosition);
+  void OnHitVfx(float damage, Vector3 attackerPosition);
 }
