+++ sim/Assets/ShopItemAsset.cs
@@ -1,3 +1,4 @@
+using xpTURN.Klotho.Deterministic.Math;
 using xpTURN.Klotho.ECS;
 using xpTURN.Klotho.Serialization;
 
@@ -7,5 +8,5 @@ namespace Meesles.Avalon.Sim.Assets;
 [KlothoDataAsset(AssetIds.TypeIds.ShopItem)]
 public partial class ShopItemAsset : IDataAsset {
   [KlothoOrder(0)] public int Cost;
-  [KlothoOrder(1)] public int AttackBonus;
+  [KlothoOrder(1)] public FP64 AttackBonus;
 }
