+++ sim/Events/AttackHitEvent.cs
@@ -8,7 +8,7 @@ namespace Meesles.Avalon.Sim;
 public partial class AttackHitEvent : SimulationEvent {
   [KlothoOrder(0)] public int AttackerUnitId;
   [KlothoOrder(1)] public int TargetUnitId;
-  [KlothoOrder(2)] public int Damage;
+  [KlothoOrder(2)] public FP64 Damage;
   [KlothoOrder(3)] public FPVector3 AttackerPosition;
   [KlothoOrder(4)] public FPVector3 TargetPosition;
 }
